Centralize RTK Query endpoints to simplify service architecture
Previously every service file contained an injectEndpoints() function
that declared its own endpoints and returned a private api object. This
meant endpoint definitions were scattered across 16 files, each with its
own boilerplate, and the relationship between endpoints and the shared
Redux store was implicit and hard to follow.

This change moves all endpoint declarations into eight static builder
files composed directly in createBackendApi(). Services no longer inject
their own endpoints — they just receive BackendApi and call
this.backendApi.endpoints.* directly. The result is that all endpoints
are visible in one place, the services themselves shrink to just
business logic, and there is no longer any question about which api
object owns which endpoint.

Two onCacheEntryAdded handlers invalidated tags via api.util, which
cannot work in static declarations because api doesn't exist yet at
construction time. Those invalidations (Diff on worktree_changes,
WorkspaceRules on rule-updates) move to explicit layout listeners and
are combined with the existing hunk-assignment-update listener via
mergeUnlisten. This is the only behavioral difference.
