+++ apps/desktop/src/lib/hunks/diffService.svelte.ts
@@ -1,9 +1,7 @@
-import { invalidatesList, providesList, ReduxTag } from "$lib/state/tags";
 import { InjectionToken } from "@gitbutler/core/context";
 import { isDefined } from "@gitbutler/ui/utils/typeguards";
 import type { TreeChange } from "$lib/hunks/change";
 import type { UnifiedDiff } from "$lib/hunks/diff";
-import type { HunkAssignmentRequest } from "$lib/hunks/hunk";
 import type { BackendApi } from "$lib/state/clientState.svelte";
 
 export type ChangeDiff = {
@@ -14,28 +12,24 @@ export type ChangeDiff = {
 export const DIFF_SERVICE = new InjectionToken<DiffService>("DiffService");
 
 export class DiffService {
-	private api: ReturnType<typeof injectEndpoints>;
-
-	constructor(api: BackendApi) {
-		this.api = injectEndpoints(api);
-	}
+	constructor(private backendApi: BackendApi) {}
 
 	getDiff(projectId: string, change: TreeChange) {
-		return this.api.endpoints.getDiff.useQuery({ projectId, change });
+		return this.backendApi.endpoints.getDiff.useQuery({ projectId, change });
 	}
 
 	get assignHunk() {
-		return this.api.endpoints.assignHunk.mutate;
+		return this.backendApi.endpoints.assignHunk.mutate;
 	}
 
 	async fetchDiff(projectId: string, change: TreeChange) {
-		const { getDiff } = this.api.endpoints;
+		const { getDiff } = this.backendApi.endpoints;
 		return await getDiff.fetch({ projectId, change });
 	}
 
 	getChanges(projectId: string, changes: TreeChange[]) {
 		const args = changes.map((change) => ({ projectId, change }));
-		const { getDiff } = this.api.endpoints;
+		const { getDiff } = this.backendApi.endpoints;
 		return getDiff.useQueries(args, {
 			transform: (data, args): ChangeDiff => ({ path: args.change.path, diff: data }),
 		});
@@ -45,7 +39,7 @@ export class DiffService {
 		const args = changes.map((change) => ({ projectId, change }));
 		const responses = await Promise.all(
 			args.map((arg) =>
-				this.api.endpoints.getDiff.fetch(arg, {
+				this.backendApi.endpoints.getDiff.fetch(arg, {
 					transform: (diff, args) => ({
 						path: args.change.path,
 						diff,
@@ -55,24 +49,4 @@ export class DiffService {
 		);
 		return responses.filter(isDefined);
 	}
-}
-function injectEndpoints(api: BackendApi) {
-	return api.injectEndpoints({
-		endpoints: (build) => ({
-			getDiff: build.query<UnifiedDiff | null, { projectId: string; change: TreeChange }>({
-				extraOptions: { command: "tree_change_diffs" },
-				query: (args) => args,
-				providesTags: [providesList(ReduxTag.Diff)],
-			}),
-			assignHunk: build.mutation<void, { projectId: string; assignments: HunkAssignmentRequest[] }>(
-				{
-					extraOptions: {
-						command: "assign_hunk",
-					},
-					query: (args) => args,
-					invalidatesTags: [invalidatesList(ReduxTag.WorktreeChanges)],
-				},
-			),
-		}),
-	});
 }
