+++ apps/desktop/src/lib/stacks/stackService.svelte.ts
@@ -1,209 +1,60 @@
 import { getBranchNameFromRef } from "$lib/branches/branchUtils";
 import { useNewRebaseEngine } from "$lib/config/uiFeatureFlags";
-import { ConflictEntries, type ConflictEntriesObj } from "$lib/files/conflicts";
 import { sortLikeFileTree } from "$lib/files/filetreeV3";
 import { showToast } from "$lib/notifications/toasts";
+import {
+	branchDetailsSelectors,
+	changesSelectors,
+	commitSelectors,
+	selectChangesByPaths,
+	stackSelectors,
+	upstreamCommitSelectors,
+} from "$lib/stacks/stackEndpoints";
 import {
 	replaceBranchInExclusiveAction,
 	replaceBranchInStackSelection,
 	updateStaleProjectState,
-	updateStaleStackState,
 } from "$lib/stacks/staleStateUpdaters";
-import { hasBackendExtra } from "$lib/state/backendQuery";
-import { createSelectByIds, createSelectNth } from "$lib/state/customSelectors";
-import {
-	invalidatesItem,
-	invalidatesList,
-	providesItem,
-	providesList,
-	ReduxTag,
-} from "$lib/state/tags";
+import { invalidatesItem, invalidatesList, ReduxTag } from "$lib/state/tags";
 import { type UiState } from "$lib/state/uiState.svelte";
 import { InjectionToken } from "@gitbutler/core/context";
 import { reactive } from "@gitbutler/shared/reactiveUtils.svelte";
 import { isDefined } from "@gitbutler/ui/utils/typeguards";
-import { createEntityAdapter, type EntityState } from "@reduxjs/toolkit";
 import { get } from "svelte/store";
-import type { StackOrder } from "$lib/branches/branch";
-import type { Commit, CommitDetails, UpstreamCommit } from "$lib/branches/v3";
-import type { MoveCommitIllegalAction } from "$lib/commits/commit";
 import type { ReduxError } from "$lib/error/reduxError";
 import type { DefaultForgeFactory } from "$lib/forge/forgeFactory.svelte";
-import type { TreeChange, TreeChanges, TreeStats } from "$lib/hunks/change";
-import type { DiffSpec } from "$lib/hunks/hunk";
-import type {
-	BranchDetails,
-	Stack,
-	StackDetails,
-	CreateRefRequest,
-	InteractiveIntegrationStep,
-	CreateBranchFromBranchOutcome,
-	MoveBranchResult,
-	GerritPushFlag,
-} from "$lib/stacks/stack";
 import type { AppDispatch, BackendApi } from "$lib/state/clientState.svelte";
-import type { RejectionReason } from "$lib/state/uiState.svelte";
 import type { HunkAssignment } from "@gitbutler/core/api";
 
-type BranchParams = {
-	name?: string;
-	order?: number;
+export type {
+	BranchParams,
+	BranchPushResult,
+	CreateCommitOutcome,
+	CreateCommitRequest,
+	CreateCommitRequestWorktreeChanges,
+	RejectionReason,
+	RelativeTo,
+	SeriesIntegrationStrategy,
+} from "$lib/stacks/stackEndpoints";
+export { REJECTTION_REASONS } from "$lib/stacks/stackEndpoints";
+
+const PUSH_ERROR_REASONS: Record<string, string> = {
+	["errors.git.authentication"]: "an authentication failure",
+	["errors.git.force_push_protection"]: "force push protection",
 };
-
-export type CreateCommitRequest = {
-	stackId: string;
-	message: string;
-	/** Undefined means that the backend will infer the parent to be the current head of stackBranchName */
-	parentId: string | undefined;
-	stackBranchName: string;
-	worktreeChanges: DiffSpec[];
-};
-
-export type CreateCommitRequestWorktreeChanges = DiffSpec;
-
-type StackAction = "push";
-
-type StackErrorInfo = {
-	title: string;
-	codeInfo: Record<string, string>;
-	defaultInfo: string;
-};
-
-const ERROR_INFO: Record<StackAction, StackErrorInfo> = {
-	push: {
-		title: "Git push failed",
-		codeInfo: {
-			["errors.git.authentication"]: "an authentication failure",
-			["errors.git.force_push_protection"]: "force push protection",
-		},
-		defaultInfo: "an unforeseen error",
-	},
-};
-
-function surfaceStackError(action: StackAction, errorCode: string, errorMessage: string): boolean {
-	const reason = ERROR_INFO[action].codeInfo[errorCode] ?? ERROR_INFO[action].defaultInfo;
-	const title = ERROR_INFO[action].title;
-	switch (action) {
-		case "push": {
-			if (errorCode === "errors.git.force_push_protection") {
-				return false;
-			}
-
-			showToast({
-				title,
-				message: `
-Your branch cannot be pushed due to ${reason}.
-
-Please check our [documentation](https://docs.gitbutler.com/troubleshooting/fetch-push)
-on fetching and pushing for ways to resolve the problem.
-			`.trim(),
-				error: errorMessage,
-				style: "danger",
-			});
-
-			return true;
-		}
-	}
-}
-
-export type SeriesIntegrationStrategy = {
-	type: "merge" | "rebase";
-};
-
-export interface BranchPushResult {
-	/**
-	 * The list of pushed branches and their corresponding remote refnames.
-	 */
-	branchToRemote: [string, string][];
-	/**
-	 * The name of the remote to which the branches were pushed.
-	 */
-	remote: string;
-}
-
-/**
- * All possible reasons for a commit to be rejected.
- *
- * This is used to display a message to the user when a commit fails.
- * @note - This reasons are in order of priority, from most to least important!
- */
-export const REJECTTION_REASONS = [
-	"workspaceMergeConflict",
-	"cherryPickMergeConflict",
-	"noEffectiveChanges",
-	"worktreeFileMissingForObjectConversion",
-	"fileToLargeOrBinary",
-	"pathNotFoundInBaseTree",
-	"unsupportedDirectoryEntry",
-	"unsupportedTreeEntry",
-	"missingDiffSpecAssociation",
-] as const;
-
-// Canonical definition lives in state/uiState.svelte.ts to avoid circular imports.
-export type { RejectionReason };
-
-type ReplacedCommit = [string, string];
-
-type BackendCreateCommitOutcome = {
-	newCommit?: string | null;
-	pathsToRejectedChanges: [RejectionReason, string][];
-	replacedCommits: Record<string, string>;
-};
-
-export type CreateCommitOutcome = {
-	newCommit: string | null;
-	pathsToRejectedChanges: [RejectionReason, string][];
-	commitMapping: ReplacedCommit[];
-};
-
-type BackendCommitRewordResult = {
-	newCommit: string;
-	replacedCommits: Record<string, string>;
-};
-
-type BackendCommitInsertBlankResult = {
-	newCommit: string;
-	replacedCommits: Record<string, string>;
-};
-
-type BackendMoveChangesResult = {
-	replacedCommits: Record<string, string>;
-};
-
-function normalizeCreateCommitOutcome(response: BackendCreateCommitOutcome): CreateCommitOutcome {
-	return {
-		newCommit: response.newCommit ?? null,
-		pathsToRejectedChanges: response.pathsToRejectedChanges,
-		commitMapping: Object.entries(response.replacedCommits),
-	};
-}
-
-export type RelativeTo =
-	| {
-			type: "commit";
-			subject: string;
-	  }
-	| {
-			type: "reference";
-			subject: string;
-	  };
 
 export const STACK_SERVICE = new InjectionToken<StackService>("StackService");
 
 export class StackService {
-	private api: ReturnType<typeof injectEndpoints>;
-
 	constructor(
-		backendApi: BackendApi,
+		private backendApi: BackendApi,
 		private dispatch: AppDispatch,
 		private forgeFactory: DefaultForgeFactory,
 		private uiState: UiState,
-	) {
-		this.api = injectEndpoints(backendApi, uiState);
-	}
+	) {}
 
 	stacks(projectId: string) {
-		return this.api.endpoints.stacks.useQuery(
+		return this.backendApi.endpoints.stacks.useQuery(
 			{ projectId },
 			{
 				transform: (stacks) => stackSelectors.selectAll(stacks),
@@ -212,7 +63,7 @@ export class StackService {
 	}
 
 	async fetchStacks(projectId: string) {
-		return await this.api.endpoints.stacks.fetch(
+		return await this.backendApi.endpoints.stacks.fetch(
 			{ projectId },
 			{
 				transform: (stacks) => stackSelectors.selectAll(stacks),
@@ -221,7 +72,7 @@ export class StackService {
 	}
 
 	stackAt(projectId: string, index: number) {
-		return this.api.endpoints.stacks.useQuery(
+		return this.backendApi.endpoints.stacks.useQuery(
 			{ projectId },
 			{
 				transform: (stacks) => stackSelectors.selectNth(stacks, index),
@@ -230,7 +81,7 @@ export class StackService {
 	}
 
 	stackById(projectId: string, id: string) {
-		return this.api.endpoints.stacks.useQuery(
+		return this.backendApi.endpoints.stacks.useQuery(
 			{ projectId },
 			{
 				transform: (stacks) => stackSelectors.selectById(stacks, id) ?? null,
@@ -239,7 +90,7 @@ export class StackService {
 	}
 
 	allStackById(projectId: string, id: string) {
-		return this.api.endpoints.stacks.useQuery(
+		return this.backendApi.endpoints.stacks.useQuery(
 			{ projectId, all: true },
 			{
 				transform: (stacks) => stackSelectors.selectById(stacks, id) ?? null,
@@ -249,7 +100,7 @@ export class StackService {
 
 	defaultBranch(projectId: string, stackId?: string) {
 		if (!stackId) return null;
-		return this.api.endpoints.stacks.useQuery(
+		return this.backendApi.endpoints.stacks.useQuery(
 			{ projectId },
 			{
 				transform: (stacks) => stackSelectors.selectById(stacks, stackId)?.heads[0]?.name ?? null,
@@ -257,15 +108,8 @@ export class StackService {
 		);
 	}
 
-	stackInfo(projectId: string, stackId: string) {
-		return this.api.endpoints.stackDetails.useQuery(
-			{ projectId, stackId },
-			{ transform: ({ stackInfo }) => stackInfo },
-		);
-	}
-
 	branchDetails(projectId: string, stackId: string | undefined, branchName?: string) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) => {
@@ -278,19 +122,19 @@ export class StackService {
 	}
 
 	get newStack() {
-		return this.api.endpoints.createStack.useMutation();
+		return this.backendApi.endpoints.createStack.useMutation();
 	}
 
 	get newStackMutation() {
-		return this.api.endpoints.createStack.mutate;
+		return this.backendApi.endpoints.createStack.mutate;
 	}
 
 	get updateStackOrder() {
-		return this.api.endpoints.updateStackOrder.mutate;
+		return this.backendApi.endpoints.updateStackOrder.mutate;
 	}
 
 	branches(projectId: string, stackId?: string) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) => branchDetailsSelectors.selectAll(branchDetails),
@@ -299,7 +143,7 @@ export class StackService {
 	}
 
 	branchAt(projectId: string, stackId: string | undefined, index: number) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ stackInfo }) => stackInfo.branchDetails[index],
@@ -309,50 +153,46 @@ export class StackService {
 
 	/** Returns the parent of the branch specified by the provided name */
 	branchParentByName(projectId: string, stackId: string | undefined, name: string) {
-		return this.api.endpoints.stackDetails.useQuery(
-			{ projectId, stackId },
-			{
-				transform: ({ stackInfo, branchDetails }) => {
-					const ids = stackInfo.branchDetails.map((branch) => branch.name);
-					const currentId = ids.findIndex((id) => id === name);
-					if (currentId === -1) return;
-
-					const parentId = ids[currentId + 1];
-					if (!parentId) return;
-
-					return branchDetailsSelectors.selectById(branchDetails, parentId);
-				},
-			},
-		);
+		return this.branchRelativeByName(projectId, stackId, name, 1);
 	}
+
 	/** Returns the child of the branch specified by the provided name */
 	branchChildByName(projectId: string, stackId: string | undefined, name: string) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.branchRelativeByName(projectId, stackId, name, -1);
+	}
+
+	private branchRelativeByName(
+		projectId: string,
+		stackId: string | undefined,
+		name: string,
+		offset: number,
+	) {
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ stackInfo, branchDetails }) => {
-					const ids = stackInfo.branchDetails.map((branch) => branch.name);
-					const currentId = ids.findIndex((id) => id === name);
-					if (currentId === -1) return;
+					const names = stackInfo.branchDetails.map((branch) => branch.name);
+					const index = names.indexOf(name);
+					if (index === -1) return;
 
-					const childId = ids[currentId - 1];
-					if (!childId) return;
+					const relativeName = names[index + offset];
+					if (!relativeName) return;
 
-					return branchDetailsSelectors.selectById(branchDetails, childId);
+					return branchDetailsSelectors.selectById(branchDetails, relativeName);
 				},
 			},
 		);
 	}
 
 	branchByName(projectId: string, stackId: string | undefined, name: string) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{ transform: ({ branchDetails }) => branchDetailsSelectors.selectById(branchDetails, name) },
 		);
 	}
 
 	commits(projectId: string, stackId: string | undefined, branchName: string) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) =>
@@ -362,7 +202,7 @@ export class StackService {
 	}
 
 	fetchCommits(projectId: string, stackId: string | undefined, branchName: string) {
-		return this.api.endpoints.stackDetails.fetch(
+		return this.backendApi.endpoints.stackDetails.fetch(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) =>
@@ -372,7 +212,7 @@ export class StackService {
 	}
 
 	async fetchStackById(projectId: string, stackId: string) {
-		return await this.api.endpoints.stacks.fetch(
+		return await this.backendApi.endpoints.stacks.fetch(
 			{ projectId },
 			{
 				transform: (stacks) => stackSelectors.selectById(stacks, stackId),
@@ -381,7 +221,7 @@ export class StackService {
 	}
 
 	async fetchBranches(projectId: string, stackId: string) {
-		return await this.api.endpoints.stackDetails.fetch(
+		return await this.backendApi.endpoints.stackDetails.fetch(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) => branchDetailsSelectors.selectAll(branchDetails),
@@ -390,7 +230,7 @@ export class StackService {
 	}
 
 	commitAt(projectId: string, stackId: string | undefined, branchName: string, index: number) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) =>
@@ -404,7 +244,7 @@ export class StackService {
 		const stackIds = $derived(stacks.response?.map((s) => s.id).filter(isDefined) || []);
 		const args = $derived(stackIds?.map((stackId) => ({ projectId, stackId })));
 		const details = $derived(
-			this.api.endpoints.stackDetails.useQueries(args, {
+			this.backendApi.endpoints.stackDetails.useQueries(args, {
 				transform: ({ commits, stackInfo }) => ({
 					commits: commitSelectors.selectAll(commits),
 					branches: stackInfo.branchDetails.map((b) => b.name),
@@ -435,7 +275,7 @@ export class StackService {
 	}
 
 	commitById(projectId: string, stackId: string | undefined, commitId: string) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ commits, upstreamCommits }) =>
@@ -446,7 +286,7 @@ export class StackService {
 	}
 
 	commitsByIds(projectId: string, stackId: string | undefined, commitIds: string[]) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ commits, upstreamCommits }) => {
@@ -463,7 +303,7 @@ export class StackService {
 	}
 
 	fetchCommitById(projectId: string, stackId: string, commitId: string) {
-		return this.api.endpoints.stackDetails.fetch(
+		return this.backendApi.endpoints.stackDetails.fetch(
 			{ projectId, stackId },
 			{
 				transform: ({ commits, upstreamCommits }) =>
@@ -474,7 +314,7 @@ export class StackService {
 	}
 
 	fetchCommitsByIds(projectId: string, stackId: string, commitIds: string[]) {
-		return this.api.endpoints.stackDetails.fetch(
+		return this.backendApi.endpoints.stackDetails.fetch(
 			{ projectId, stackId },
 			{
 				transform: ({ commits, upstreamCommits }) => {
@@ -491,7 +331,7 @@ export class StackService {
 	}
 
 	upstreamCommits(projectId: string, stackId: string | undefined, branchName: string) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) =>
@@ -501,7 +341,7 @@ export class StackService {
 	}
 
 	upstreamCommitAt(projectId: string, stackId: string, branchName: string, index: number) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) =>
@@ -512,7 +352,7 @@ export class StackService {
 	}
 
 	fetchUpstreamCommitById(projectId: string, stackId: string, commitId: string) {
-		return this.api.endpoints.stackDetails.fetch(
+		return this.backendApi.endpoints.stackDetails.fetch(
 			{ projectId, stackId },
 			{
 				transform: ({ upstreamCommits }) =>
@@ -522,7 +362,7 @@ export class StackService {
 	}
 
 	get pushStack() {
-		return this.api.endpoints.pushStack.useMutation({
+		return this.backendApi.endpoints.pushStack.useMutation({
 			sideEffect: (result, _) => {
 				// Timeout to accommodate eventual consistency.
 				setTimeout(() => {
@@ -542,10 +382,16 @@ export class StackService {
 			},
 			onError: (commandError: ReduxError) => {
 				const { code, message } = commandError;
-				const handled = surfaceStackError("push", code ?? "", message);
-				if (!handled && code === "errors.git.force_push_protection") {
+				if (code === "errors.git.force_push_protection") {
 					throw commandError;
 				}
+				const reason = PUSH_ERROR_REASONS[code ?? ""] ?? "an unforeseen error";
+				showToast({
+					title: "Git push failed",
+					message: `Your branch cannot be pushed due to ${reason}.\n\nPlease check our [documentation](https://docs.gitbutler.com/troubleshooting/fetch-push)\non fetching and pushing for ways to resolve the problem.`,
+					error: message,
+					style: "danger",
+				});
 			},
 			throwSilentError: true,
 		});
@@ -553,17 +399,17 @@ export class StackService {
 
 	createCommit() {
 		if (get(useNewRebaseEngine)) {
-			return this.api.endpoints.commitCreate.useMutation();
+			return this.backendApi.endpoints.commitCreate.useMutation();
 		} else {
-			return this.api.endpoints.legacyCreateCommit.useMutation();
+			return this.backendApi.endpoints.legacyCreateCommit.useMutation();
 		}
 	}
 
 	get createCommitMutation() {
 		if (get(useNewRebaseEngine)) {
-			return this.api.endpoints.commitCreate.mutate;
+			return this.backendApi.endpoints.commitCreate.mutate;
 		} else {
-			return this.api.endpoints.legacyCreateCommit.mutate;
+			return this.backendApi.endpoints.legacyCreateCommit.mutate;
 		}
 	}
 
@@ -572,7 +418,7 @@ export class StackService {
 			projectId,
 			commitId,
 		}));
-		return this.api.endpoints.commitDetails.useQueries(params, {
+		return this.backendApi.endpoints.commitDetails.useQueries(params, {
 			transform: (results) => {
 				return results.changes.ids;
 			},
@@ -580,7 +426,7 @@ export class StackService {
 	}
 
 	commitChanges(projectId: string, commitId: string) {
-		return this.api.endpoints.commitDetails.useQuery(
+		return this.backendApi.endpoints.commitDetails.useQuery(
 			{ projectId, commitId },
 			{
 				transform: (result) => ({
@@ -593,7 +439,7 @@ export class StackService {
 	}
 
 	fetchCommitChanges(projectId: string, commitId: string) {
-		return this.api.endpoints.commitDetails.fetch(
+		return this.backendApi.endpoints.commitDetails.fetch(
 			{ projectId, commitId },
 			{
 				transform: (result) => ({
@@ -606,14 +452,14 @@ export class StackService {
 	}
 
 	commitChange(projectId: string, commitId: string, path: string) {
-		return this.api.endpoints.commitDetails.useQuery(
+		return this.backendApi.endpoints.commitDetails.useQuery(
 			{ projectId, commitId },
 			{ transform: (result) => changesSelectors.selectById(result.changes, path) },
 		);
 	}
 
 	async commitChangesByPaths(projectId: string, commitId: string, paths: string[]) {
-		const result = await this.api.endpoints.commitDetails.fetch(
+		const result = await this.backendApi.endpoints.commitDetails.fetch(
 			{ projectId, commitId },
 			{ transform: (result) => selectChangesByPaths(result.changes, paths) },
 		);
@@ -621,14 +467,14 @@ export class StackService {
 	}
 
 	commitDetails(projectId: string, commitId: string) {
-		return this.api.endpoints.commitDetails.useQuery(
+		return this.backendApi.endpoints.commitDetails.useQuery(
 			{ projectId, commitId },
 			{ transform: (result) => result.details },
 		);
 	}
 
 	fetchCommitDetails(projectId: string, commitId: string) {
-		return this.api.endpoints.commitDetails.fetch(
+		return this.backendApi.endpoints.commitDetails.fetch(
 			{ projectId, commitId },
 			{ transform: (result) => result.details },
 		);
@@ -638,7 +484,7 @@ export class StackService {
 	 * Gets the changes for a given branch.
 	 */
 	branchChanges(args: { projectId: string; stackId?: string; branch: string }) {
-		return this.api.endpoints.branchChanges.useQuery(
+		return this.backendApi.endpoints.branchChanges.useQuery(
 			{
 				projectId: args.projectId,
 				branch: args.branch,
@@ -654,7 +500,7 @@ export class StackService {
 	}
 
 	branchChange(args: { projectId: string; stackId?: string; branch: string; path: string }) {
-		return this.api.endpoints.branchChanges.useQuery(
+		return this.backendApi.endpoints.branchChanges.useQuery(
 			{
 				projectId: args.projectId,
 				stackId: args.stackId,
@@ -670,7 +516,7 @@ export class StackService {
 		branch: string;
 		paths: string[];
 	}) {
-		const result = await this.api.endpoints.branchChanges.fetch(
+		const result = await this.backendApi.endpoints.branchChanges.fetch(
 			{
 				projectId: args.projectId,
 				stackId: args.stackId,
@@ -683,14 +529,14 @@ export class StackService {
 
 	get updateCommitMessage() {
 		if (get(useNewRebaseEngine)) {
-			return this.api.endpoints.updateCommitMessage.useMutation();
+			return this.backendApi.endpoints.updateCommitMessage.useMutation();
 		} else {
-			return this.api.endpoints.legacyUpdateCommitMessage.useMutation();
+			return this.backendApi.endpoints.legacyUpdateCommitMessage.useMutation();
 		}
 	}
 
 	get newBranch() {
-		return this.api.endpoints.newBranch.useMutation();
+		return this.backendApi.endpoints.newBranch.useMutation();
 	}
 
 	async uncommit(args: {
@@ -699,7 +545,7 @@ export class StackService {
 		branchName: string;
 		commitId: string;
 	}) {
-		const result = await this.api.endpoints.uncommit.mutate(args);
+		const result = await this.backendApi.endpoints.uncommit.mutate(args);
 		const selection = this.uiState.lane(args.stackId).selection;
 		if (args.commitId === selection.current?.commitId) {
 			selection.set(undefined);
@@ -708,43 +554,43 @@ export class StackService {
 	}
 
 	get insertBlankCommit() {
-		return this.api.endpoints.insertBlankCommit;
+		return this.backendApi.endpoints.insertBlankCommit;
 	}
 
 	get unapply() {
-		return this.api.endpoints.unapply.mutate;
+		return this.backendApi.endpoints.unapply.mutate;
 	}
 
 	get discardChanges() {
-		return this.api.endpoints.discardChanges.mutate;
+		return this.backendApi.endpoints.discardChanges.mutate;
 	}
 
 	get moveChangesBetweenCommits() {
 		if (get(useNewRebaseEngine)) {
-			return this.api.endpoints.commitMoveChangesBetween.mutate;
+			return this.backendApi.endpoints.commitMoveChangesBetween.mutate;
 		} else {
-			return this.api.endpoints.legacyMoveChangesBetweenCommits.mutate;
+			return this.backendApi.endpoints.legacyMoveChangesBetweenCommits.mutate;
 		}
 	}
 
 	get uncommitChanges() {
 		if (get(useNewRebaseEngine)) {
-			return this.api.endpoints.commitUncommitChanges.mutate;
+			return this.backendApi.endpoints.commitUncommitChanges.mutate;
 		} else {
-			return this.api.endpoints.legacyUncommitChanges.mutate;
+			return this.backendApi.endpoints.legacyUncommitChanges.mutate;
 		}
 	}
 
 	get stashIntoBranch() {
-		return this.api.endpoints.stashIntoBranch.mutate;
+		return this.backendApi.endpoints.stashIntoBranch.mutate;
 	}
 
 	get updateBranchPrNumber() {
-		return this.api.endpoints.updateBranchPrNumber.mutate;
+		return this.backendApi.endpoints.updateBranchPrNumber.mutate;
 	}
 
 	get updateBranchName() {
-		return this.api.endpoints.updateBranchName.useMutation({
+		return this.backendApi.endpoints.updateBranchName.useMutation({
 			sideEffect: (_, args) => {
 				// Immediately update the selection and the exclusive action.
 				const laneState = this.uiState.lane(args.laneId);
@@ -782,31 +628,31 @@ export class StackService {
 	}
 
 	get removeBranch() {
-		return this.api.endpoints.removeBranch.useMutation();
+		return this.backendApi.endpoints.removeBranch.useMutation();
 	}
 
 	get reorderStack() {
-		return this.api.endpoints.reorderStack.mutate;
+		return this.backendApi.endpoints.reorderStack.mutate;
 	}
 
 	get moveCommit() {
-		return this.api.endpoints.moveCommit.mutate;
+		return this.backendApi.endpoints.moveCommit.mutate;
 	}
 
 	get moveBranch() {
-		return this.api.endpoints.moveBranch.mutate;
+		return this.backendApi.endpoints.moveBranch.mutate;
 	}
 
 	get tearOffBranch() {
-		return this.api.endpoints.tearOffBranch.mutate;
+		return this.backendApi.endpoints.tearOffBranch.mutate;
 	}
 
 	get integrateUpstreamCommits() {
-		return this.api.endpoints.integrateUpstreamCommits.useMutation();
+		return this.backendApi.endpoints.integrateUpstreamCommits.useMutation();
 	}
 
 	initialIntegrationSteps(projectId: string, stackId: string | undefined, branchName: string) {
-		return this.api.endpoints.getInitialIntegrationSteps.useQuery({
+		return this.backendApi.endpoints.getInitialIntegrationSteps.useQuery({
 			projectId,
 			stackId,
 			branchName,
@@ -814,34 +660,34 @@ export class StackService {
 	}
 
 	get integrateBranchWithSteps() {
-		return this.api.endpoints.integrateBranchWithSteps.useMutation();
+		return this.backendApi.endpoints.integrateBranchWithSteps.useMutation();
 	}
 
 	get createVirtualBranchFromBranch() {
-		return this.api.endpoints.createVirtualBranchFromBranch.mutate;
+		return this.backendApi.endpoints.createVirtualBranchFromBranch.mutate;
 	}
 
 	get deleteLocalBranch() {
-		return this.api.endpoints.deleteLocalBranch.mutate;
+		return this.backendApi.endpoints.deleteLocalBranch.mutate;
 	}
 
 	get squashCommits() {
-		return this.api.endpoints.squashCommits.mutate;
+		return this.backendApi.endpoints.squashCommits.mutate;
 	}
 
 	get amendCommit() {
 		if (get(useNewRebaseEngine)) {
-			return this.api.endpoints.commitAmend.useMutation();
+			return this.backendApi.endpoints.commitAmend.useMutation();
 		} else {
-			return this.api.endpoints.legacyAmendCommit.useMutation();
+			return this.backendApi.endpoints.legacyAmendCommit.useMutation();
 		}
 	}
 
 	get amendCommitMutation() {
 		if (get(useNewRebaseEngine)) {
-			return this.api.endpoints.commitAmend.mutate;
+			return this.backendApi.endpoints.commitAmend.mutate;
 		} else {
-			return this.api.endpoints.legacyAmendCommit.mutate;
+			return this.backendApi.endpoints.legacyAmendCommit.mutate;
 		}
 	}
 
@@ -855,7 +701,7 @@ export class StackService {
 		stackId: string;
 		branchName: string;
 	}) {
-		const allCommits = await this.api.endpoints.stackDetails.fetch(
+		const allCommits = await this.backendApi.endpoints.stackDetails.fetch(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) =>
@@ -880,15 +726,18 @@ export class StackService {
 	}
 
 	newBranchName(projectId: string) {
-		return this.api.endpoints.newBranchName.useQuery({ projectId }, { forceRefetch: true });
+		return this.backendApi.endpoints.newBranchName.useQuery({ projectId }, { forceRefetch: true });
 	}
 
 	async fetchNewBranchName(projectId: string) {
-		return await this.api.endpoints.newBranchName.fetch({ projectId }, { forceRefetch: true });
+		return await this.backendApi.endpoints.newBranchName.fetch(
+			{ projectId },
+			{ forceRefetch: true },
+		);
 	}
 
 	isBranchConflicted(projectId: string, stackId: string, branchName: string) {
-		return this.api.endpoints.stackDetails.useQuery(
+		return this.backendApi.endpoints.stackDetails.useQuery(
 			{ projectId, stackId },
 			{
 				transform: ({ branchDetails }) => {
@@ -900,7 +749,10 @@ export class StackService {
 	}
 
 	async normalizeBranchName(name: string) {
-		return await this.api.endpoints.normalizeBranchName.fetch({ name }, { forceRefetch: true });
+		return await this.backendApi.endpoints.normalizeBranchName.fetch(
+			{ name },
+			{ forceRefetch: true },
+		);
 	}
 
 	/**
@@ -908,14 +760,14 @@ export class StackService {
 	 * a stacking context. You almost certainly want `stackDetails`
 	 */
 	unstackedBranchDetails(projectId: string, branchName: string, remote?: string) {
-		return this.api.endpoints.unstackedBranchDetails.useQuery(
+		return this.backendApi.endpoints.unstackedBranchDetails.useQuery(
 			{ projectId, branchName, remote },
 			{ transform: (result) => result.branchDetails },
 		);
 	}
 
 	unstackedCommits(projectId: string, branchName: string, remote?: string) {
-		return this.api.endpoints.unstackedBranchDetails.useQuery(
+		return this.backendApi.endpoints.unstackedBranchDetails.useQuery(
 			{ projectId, branchName, remote },
 			{
 				transform: (data) => commitSelectors.selectAll(data.commits),
@@ -924,7 +776,7 @@ export class StackService {
 	}
 
 	async fetchUnstackedCommits(projectId: string, branchName: string, remote?: string) {
-		return await this.api.endpoints.unstackedBranchDetails.fetch(
+		return await this.backendApi.endpoints.unstackedBranchDetails.fetch(
 			{ projectId, branchName, remote },
 			{
 				transform: (data) => commitSelectors.selectAll(data.commits),
@@ -933,14 +785,14 @@ export class StackService {
 	}
 
 	unstackedCommitById(projectId: string, branchName: string, commitId: string, remote?: string) {
-		return this.api.endpoints.unstackedBranchDetails.useQuery(
+		return this.backendApi.endpoints.unstackedBranchDetails.useQuery(
 			{ projectId, branchName, remote },
 			{ transform: ({ commits }) => commitSelectors.selectById(commits, commitId) },
 		);
 	}
 
 	async targetCommits(projectId: string, lastCommitId: string | undefined, pageSize: number) {
-		return await this.api.endpoints.targetCommits.fetch(
+		return await this.backendApi.endpoints.targetCommits.fetch(
 			{ projectId, lastCommitId, pageSize },
 			{
 				forceRefetch: true,
@@ -950,20 +802,20 @@ export class StackService {
 	}
 
 	get splitBranch() {
-		return this.api.endpoints.splitBranch.useMutation();
+		return this.backendApi.endpoints.splitBranch.useMutation();
 	}
 
 	get splitBranchMutation() {
-		return this.api.endpoints.splitBranch.mutate;
+		return this.backendApi.endpoints.splitBranch.mutate;
 	}
 
-	get splitBrancIntoDependentBranch() {
-		return this.api.endpoints.splitBranchIntoDependentBranch.useMutation();
+	get splitBranchIntoDependentBranch() {
+		return this.backendApi.endpoints.splitBranchIntoDependentBranch.useMutation();
 	}
 
 	invalidateStacksAndDetails() {
 		this.dispatch(
-			this.api.util.invalidateTags([
+			this.backendApi.util.invalidateTags([
 				invalidatesList(ReduxTag.Stacks),
 				invalidatesList(ReduxTag.StackDetails),
 			]),
@@ -971,11 +823,11 @@ export class StackService {
 	}
 
 	templates(projectId: string, forgeName: string) {
-		return this.api.endpoints.templates.useQuery({ projectId, forge: forgeName });
+		return this.backendApi.endpoints.templates.useQuery({ projectId, forge: forgeName });
 	}
 
 	async template(projectId: string, forgeName: string, relativePath: string) {
-		return await this.api.endpoints.template.fetch({
+		return await this.backendApi.endpoints.template.fetch({
 			relativePath,
 			projectId,
 			forge: forgeName,
@@ -983,954 +835,14 @@ export class StackService {
 	}
 
 	get createReference() {
-		return this.api.endpoints.createReference.useMutation();
+		return this.backendApi.endpoints.createReference.useMutation();
 	}
 
 	get absorb() {
-		return this.api.endpoints.absorb.useMutation();
+		return this.backendApi.endpoints.absorb.useMutation();
 	}
 
 	async fetchAbsorbPlan(projectId: string, target: HunkAssignment.AbsorptionTarget) {
-		return await this.api.endpoints.absorbPlan.fetch({ projectId, target });
+		return await this.backendApi.endpoints.absorbPlan.fetch({ projectId, target });
 	}
 }
-
-function transformStacksResponse(response: Stack[]) {
-	// response.forEach((stack) => {
-	// 	// To keep it simple, what's cast as `Stack` is actually `StackOpt`
-	// 	// (as returned by the backend).
-	// 	// So here we cast it back and stop any optional stack-id in its tracks
-	// 	// until the code can actually cope with it.
-	// 	const stackOpt = stack as StackOpt;
-	// 	if (!stackOpt.id) {
-	// 		throw new Error('BUG(opt-stack-id): cannot yet handle optional stack IDs');
-	// 	}
-	// });
-	return stackAdapter.addMany(stackAdapter.getInitialState(), response);
-}
-
-function toCommitCreatePlacement(args: CreateCommitRequest): {
-	relativeTo: RelativeTo;
-	side: "above" | "below";
-} {
-	if (args.parentId) {
-		return {
-			relativeTo: {
-				type: "commit",
-				subject: args.parentId,
-			},
-			side: "above",
-		};
-	}
-
-	return {
-		relativeTo: {
-			type: "reference",
-			subject: args.stackBranchName.startsWith("refs/")
-				? args.stackBranchName
-				: `refs/heads/${args.stackBranchName}`,
-		},
-		side: "below",
-	};
-}
-
-function injectEndpoints(api: BackendApi, uiState: UiState) {
-	return api.injectEndpoints({
-		endpoints: (build) => ({
-			stacks: build.query<EntityState<Stack, string>, { projectId: string; all?: boolean }>({
-				extraOptions: { command: "stacks" },
-				query: (args) => {
-					const filter = args.all ? "All" : undefined;
-					return { projectId: args.projectId, filter };
-				},
-				providesTags: [providesList(ReduxTag.Stacks)],
-				async onCacheEntryAdded(arg, lifecycleApi) {
-					if (!hasBackendExtra(lifecycleApi.extra)) {
-						throw new Error("Redux dependency Tauri not found!");
-					}
-					// The `cacheDataLoaded` promise resolves when the result is first loaded.
-					await lifecycleApi.cacheDataLoaded;
-					const unsubscribe = lifecycleApi.extra.backend.listen(
-						`project://${arg.projectId}/hunk-assignment-update`,
-						() => {
-							lifecycleApi.dispatch(api.util.invalidateTags([invalidatesList(ReduxTag.Stacks)]));
-						},
-					);
-					// The `cacheEntryRemoved` promise resolves when the result is removed
-					await lifecycleApi.cacheEntryRemoved;
-					unsubscribe();
-				},
-				transformResponse(response: Stack[], _) {
-					return transformStacksResponse(response);
-				},
-			}),
-			createStack: build.mutation<Stack, { projectId: string; branch: BranchParams }>({
-				extraOptions: {
-					command: "create_virtual_branch",
-					actionName: "Create Stack",
-				},
-				query: (args) => args,
-				invalidatesTags: (result, _error) => [
-					invalidatesItem(ReduxTag.StackDetails, result?.id || "undefined"),
-					invalidatesList(ReduxTag.Stacks),
-					invalidatesList(ReduxTag.UpstreamIntegrationStatus),
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			updateStackOrder: build.mutation<
-				void,
-				{ projectId: string; stacks: { id: string; order: number }[] }
-			>({
-				extraOptions: {
-					command: "update_stack_order",
-					actionName: "Update Stack Order",
-				},
-				query: (args) => args,
-				// This invalidation causes the order to jump back and forth
-				// on save, and it's a bit unclear why. It's not important to
-				// reload, however, so leaving it like this for now.
-				// invalidatesTags: [invalidatesList(ReduxTag.Stacks)]
-			}),
-			stackDetails: build.query<
-				{
-					stackInfo: StackDetails;
-					branchDetails: EntityState<BranchDetails, string>;
-					commits: EntityState<Commit, string>;
-					upstreamCommits: EntityState<UpstreamCommit, string>;
-				},
-				// TODO(single-branch): stackId is actually `stackId?` in the backend to be able to query details in single-branch mode.
-				// 	  however, ideally all this goes away in favor of consuming `RefInfo` from the backend.
-				{ projectId: string; stackId?: string }
-			>({
-				extraOptions: { command: "stack_details" },
-				query: (args) => args,
-				providesTags: (_result, _error, { stackId }) => [
-					...providesItem(ReduxTag.StackDetails, stackId || "undefined"),
-				],
-				transformResponse(response: StackDetails, _, { stackId }) {
-					if (stackId) {
-						updateStaleStackState(uiState, stackId, response);
-					}
-
-					const branchDetailsEntity = branchDetailsAdapter.addMany(
-						branchDetailsAdapter.getInitialState(),
-						response.branchDetails,
-					);
-
-					// This is a list of all the commits across all branches in the stack.
-					// If you want to access the commits of a specific branch, use the
-					// `commits` property of the `BranchDetails` struct.
-					const commitsEntity = commitAdapter.addMany(
-						commitAdapter.getInitialState(),
-						response.branchDetails.flatMap((branch) => branch.commits),
-					);
-
-					// This is a list of all the upstream commits across all the branches in the stack.
-					// If you want to access the upstream commits of a specific branch, use the
-					// `upstreamCommits` property of the `BranchDetails` struct.
-					const upstreamCommitsEntity = upstreamCommitAdapter.addMany(
-						upstreamCommitAdapter.getInitialState(),
-						response.branchDetails.flatMap((branch) => branch.upstreamCommits),
-					);
-
-					return {
-						stackInfo: response,
-						branchDetails: branchDetailsEntity,
-						commits: commitsEntity,
-						upstreamCommits: upstreamCommitsEntity,
-					};
-				},
-			}),
-			/**
-			 * Note: This is specifically for looking up branches outside of
-			 * a stacking context. You almost certainly want `stackDetails`
-			 */
-			unstackedBranchDetails: build.query<
-				{
-					branchDetails: BranchDetails;
-					commits: EntityState<Commit, string>;
-					upstreamCommits: EntityState<UpstreamCommit, string>;
-				},
-				{ projectId: string; branchName: string; remote?: string }
-			>({
-				extraOptions: { command: "branch_details" },
-				query: (args) => args,
-				transformResponse(branchDetails: BranchDetails) {
-					// This is a list of all the commits across all branches in the stack.
-					// If you want to access the commits of a specific branch, use the
-					// `commits` property of the `BranchDetails` struct.
-					const commitsEntity = commitAdapter.addMany(
-						commitAdapter.getInitialState(),
-						branchDetails.commits,
-					);
-
-					// This is a list of all the upstream commits across all the branches in the stack.
-					// If you want to access the upstream commits of a specific branch, use the
-					// `upstreamCommits` property of the `BranchDetails` struct.
-					const upstreamCommitsEntity = upstreamCommitAdapter.addMany(
-						upstreamCommitAdapter.getInitialState(),
-						branchDetails.upstreamCommits,
-					);
-
-					return {
-						branchDetails,
-						commits: commitsEntity,
-						upstreamCommits: upstreamCommitsEntity,
-					};
-				},
-				providesTags: (_result, _error, { branchName }) => [
-					...providesItem(ReduxTag.BranchDetails, branchName),
-				],
-			}),
-			pushStack: build.mutation<
-				BranchPushResult,
-				{
-					projectId: string;
-					stackId: string;
-					withForce: boolean;
-					skipForcePushProtection: boolean;
-					branch: string;
-					runHooks: boolean;
-					pushOpts: GerritPushFlag[];
-				}
-			>({
-				extraOptions: {
-					command: "push_stack",
-					actionName: "Push",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesItem(ReduxTag.StackDetails, args.stackId), // Is this still needed?
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			legacyCreateCommit: build.mutation<
-				CreateCommitOutcome,
-				{ projectId: string } & CreateCommitRequest
-			>({
-				extraOptions: {
-					command: "create_commit_from_worktree_changes",
-					actionName: "Commit",
-				},
-				query: (args) => args,
-				invalidatesTags: [
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesList(ReduxTag.UpstreamIntegrationStatus),
-					invalidatesList(ReduxTag.HeadSha),
-				],
-			}),
-			commitCreate: build.mutation<
-				CreateCommitOutcome,
-				{ projectId: string } & CreateCommitRequest
-			>({
-				extraOptions: {
-					command: "commit_create",
-					actionName: "Commit",
-				},
-				query: (args) => {
-					const { relativeTo, side } = toCommitCreatePlacement(args);
-					return {
-						projectId: args.projectId,
-						relativeTo,
-						side,
-						changes: args.worktreeChanges,
-						message: args.message,
-					};
-				},
-				transformResponse: normalizeCreateCommitOutcome,
-				invalidatesTags: [
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesList(ReduxTag.UpstreamIntegrationStatus),
-					invalidatesList(ReduxTag.HeadSha),
-				],
-			}),
-			commitDetails: build.query<
-				{
-					changes: EntityState<TreeChange, string>;
-					details: Commit;
-					stats: TreeStats;
-					conflictEntries?: ConflictEntriesObj;
-				},
-				{ projectId: string; commitId: string }
-			>({
-				keepUnusedDataFor: 60, // Keep for 1 minute after last use
-				extraOptions: { command: "commit_details_with_line_stats" },
-				query: (args) => args,
-				providesTags: (_result, _error, { commitId }) => [
-					...providesItem(ReduxTag.CommitChanges, commitId),
-				],
-				transformResponse(rsp: CommitDetails) {
-					const changes = changesAdapter.addMany(changesAdapter.getInitialState(), rsp.changes);
-					const stats = rsp.stats;
-					return {
-						changes: changes,
-						details: rsp.commit,
-						stats,
-						conflictEntries: rsp.conflictEntries
-							? new ConflictEntries(
-									rsp.conflictEntries.ancestorEntries,
-									rsp.conflictEntries.ourEntries,
-									rsp.conflictEntries.theirEntries,
-								).toObj()
-							: undefined,
-					};
-				},
-			}),
-			branchChanges: build.query<
-				{ changes: EntityState<TreeChange, string>; stats: TreeStats },
-				{ projectId: string; stackId?: string; branch: string }
-			>({
-				extraOptions: { command: "branch_diff" },
-				query: (args) => args,
-				providesTags: (_result, _error, { stackId }) =>
-					stackId ? providesItem(ReduxTag.BranchChanges, stackId) : [],
-				transformResponse(rsp: TreeChanges) {
-					return {
-						changes: changesAdapter.addMany(changesAdapter.getInitialState(), rsp.changes),
-						stats: rsp.stats,
-					};
-				},
-			}),
-			legacyUpdateCommitMessage: build.mutation<
-				string,
-				{ projectId: string; stackId: string; commitId: string; message: string }
-			>({
-				extraOptions: {
-					command: "update_commit_message",
-					actionName: "Update Commit Message",
-				},
-				query: (args) => args,
-				invalidatesTags: [invalidatesList(ReduxTag.HeadSha)],
-			}),
-			updateCommitMessage: build.mutation<
-				string,
-				{ projectId: string; commitId: string; message: string }
-			>({
-				extraOptions: {
-					command: "commit_reword",
-					actionName: "Update Commit Message",
-				},
-				query: (args) => args,
-				transformResponse: (response: BackendCommitRewordResult) => response.newCommit,
-				invalidatesTags: [invalidatesList(ReduxTag.HeadSha)],
-			}),
-			newBranch: build.mutation<
-				void,
-				{ projectId: string; stackId: string; request: { targetPatch?: string; name: string } }
-			>({
-				extraOptions: {
-					command: "create_branch",
-					actionName: "Create Branch",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesItem(ReduxTag.StackDetails, args.stackId),
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			uncommit: build.mutation<void, { projectId: string; stackId: string; commitId: string }>({
-				extraOptions: {
-					command: "undo_commit",
-					actionName: "Uncommit",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesItem(ReduxTag.BranchChanges, args.stackId),
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesList(ReduxTag.HeadSha),
-				],
-			}),
-			legacyAmendCommit: build.mutation<
-				string /** Return value is the updated commit id. */,
-				{
-					projectId: string;
-					stackId: string;
-					commitId: string;
-					worktreeChanges: DiffSpec[];
-				}
-			>({
-				extraOptions: {
-					command: "amend_virtual_branch",
-					actionName: "Amend Commit",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesItem(ReduxTag.BranchChanges, args.stackId),
-					invalidatesList(ReduxTag.HeadSha),
-				],
-			}),
-			commitAmend: build.mutation<
-				string /** Return value is the updated commit id. */,
-				{
-					projectId: string;
-					commitId: string;
-					worktreeChanges: DiffSpec[];
-				}
-			>({
-				extraOptions: {
-					command: "commit_amend",
-					actionName: "Amend Commit",
-				},
-				query: ({ projectId, commitId, worktreeChanges }) => ({
-					projectId,
-					commitId,
-					changes: worktreeChanges,
-				}),
-				transformResponse: (response: BackendCreateCommitOutcome) => {
-					const normalizedResponse = normalizeCreateCommitOutcome(response);
-					if (normalizedResponse.newCommit) {
-						return normalizedResponse.newCommit;
-					}
-
-					const rejected = normalizedResponse.pathsToRejectedChanges
-						.map(([reason, path]) => `${reason}: ${path}`)
-						.join(", ");
-					const details = rejected ? ` Rejected changes: ${rejected}` : "";
-					throw new Error(`Failed to amend commit: no commit was created.${details}`);
-				},
-				invalidatesTags: [
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesList(ReduxTag.BranchChanges),
-					invalidatesList(ReduxTag.HeadSha),
-				],
-			}),
-			absorbPlan: build.query<
-				HunkAssignment.CommitAbsorption[],
-				{ projectId: string; target: HunkAssignment.AbsorptionTarget }
-			>({
-				extraOptions: { command: "absorption_plan" },
-				query: (args) => args,
-			}),
-			absorb: build.mutation<
-				number,
-				{ projectId: string; absorptionPlan: HunkAssignment.CommitAbsorption[] }
-			>({
-				extraOptions: {
-					command: "absorb",
-					actionName: "Absorb changes v2",
-				},
-				query: (args) => args,
-				invalidatesTags: [
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesList(ReduxTag.HeadSha),
-				],
-			}),
-			insertBlankCommit: build.mutation<
-				string,
-				{
-					projectId: string;
-					relativeTo: RelativeTo;
-					side: "above" | "below";
-				}
-			>({
-				extraOptions: {
-					command: "commit_insert_blank",
-					actionName: "Insert Blank Commit",
-				},
-				query: (args) => args,
-				transformResponse: (response: BackendCommitInsertBlankResult) => response.newCommit,
-				invalidatesTags: [invalidatesList(ReduxTag.HeadSha)],
-			}),
-			discardChanges: build.mutation<
-				DiffSpec[],
-				{ projectId: string; worktreeChanges: DiffSpec[] }
-			>({
-				extraOptions: {
-					command: "discard_worktree_changes",
-					actionName: "Discard Changes",
-				},
-				query: (args) => args,
-				invalidatesTags: [invalidatesList(ReduxTag.WorktreeChanges)],
-			}),
-			legacyMoveChangesBetweenCommits: build.mutation<
-				{ replacedCommits: [string, string][] },
-				{
-					projectId: string;
-					changes: DiffSpec[];
-					sourceCommitId: string;
-					sourceStackId: string;
-					destinationCommitId: string;
-					destinationStackId: string;
-				}
-			>({
-				extraOptions: {
-					command: "move_changes_between_commits",
-					actionName: "Move Changes Between Commits",
-				},
-				query: (args) => args,
-				invalidatesTags(result, _error, arg) {
-					const commitChangesTags = [arg.sourceCommitId, arg.destinationCommitId]
-						.map((id) => result?.replacedCommits.find(([oldId]) => oldId === id)?.[1])
-						.filter(isDefined)
-						.map((id) => invalidatesItem(ReduxTag.CommitChanges, id));
-					return [
-						invalidatesList(ReduxTag.HeadSha),
-						invalidatesList(ReduxTag.WorktreeChanges),
-						invalidatesItem(ReduxTag.BranchChanges, arg.sourceStackId),
-						invalidatesItem(ReduxTag.BranchChanges, arg.destinationStackId),
-						...commitChangesTags,
-					];
-				},
-			}),
-			commitMoveChangesBetween: build.mutation<
-				{
-					replacedCommits: ReplacedCommit[];
-				},
-				{
-					projectId: string;
-					changes: DiffSpec[];
-					sourceCommitId: string;
-					destinationCommitId: string;
-				}
-			>({
-				extraOptions: {
-					command: "commit_move_changes_between",
-					actionName: "Move Changes Between Commits",
-				},
-				query: (args) => args,
-				transformResponse: (a: BackendMoveChangesResult) => ({
-					replacedCommits: Object.entries(a.replacedCommits),
-				}),
-				invalidatesTags: [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesList(ReduxTag.CommitChanges),
-				],
-			}),
-			legacyUncommitChanges: build.mutation<
-				{ replacedCommits: [string, string][] },
-				{
-					projectId: string;
-					changes: DiffSpec[];
-					commitId: string;
-					stackId: string;
-					assignTo?: string;
-				}
-			>({
-				extraOptions: {
-					command: "uncommit_changes",
-					actionName: "Uncommit Changes",
-				},
-				query: (args) => args,
-				invalidatesTags(_result, _error, args) {
-					return [
-						invalidatesList(ReduxTag.HeadSha),
-						invalidatesList(ReduxTag.WorktreeChanges),
-						invalidatesItem(ReduxTag.BranchChanges, args.stackId),
-					];
-				},
-			}),
-			commitUncommitChanges: build.mutation<
-				{
-					replacedCommits: ReplacedCommit[];
-				},
-				{
-					projectId: string;
-					changes: DiffSpec[];
-					commitId: string;
-					assignTo?: string;
-				}
-			>({
-				extraOptions: {
-					command: "commit_uncommit_changes",
-					actionName: "Uncommit Changes",
-				},
-				query: (args) => args,
-				transformResponse: (a: BackendMoveChangesResult) => ({
-					replacedCommits: Object.entries(a.replacedCommits),
-				}),
-				invalidatesTags() {
-					return [
-						invalidatesList(ReduxTag.HeadSha),
-						invalidatesList(ReduxTag.WorktreeChanges),
-						invalidatesList(ReduxTag.BranchChanges),
-					];
-				},
-			}),
-			stashIntoBranch: build.mutation<
-				DiffSpec[],
-				{ projectId: string; branchName: string; worktreeChanges: DiffSpec[] }
-			>({
-				extraOptions: {
-					command: "stash_into_branch",
-					actionName: "Stash Changes",
-				},
-				query: (args) => args,
-				invalidatesTags: [
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			unapply: build.mutation<void, { projectId: string; stackId: string }>({
-				extraOptions: {
-					command: "unapply_stack",
-					actionName: "Unapply Stack",
-				},
-				query: (args) => args,
-				invalidatesTags: [
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			// TODO: Why is this not part of the regular update call?
-			updateBranchPrNumber: build.mutation<
-				void,
-				{
-					projectId: string;
-					stackId: string;
-					branchName: string;
-					prNumber?: number;
-				}
-			>({
-				extraOptions: {
-					command: "update_branch_pr_number",
-					actionName: "Update Branch PR Number",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesItem(ReduxTag.StackDetails, args.stackId), // This probably is still needed
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			updateBranchName: build.mutation<
-				void,
-				{
-					projectId: string;
-					stackId?: string;
-					laneId: string;
-					branchName: string;
-					newName: string;
-				}
-			>({
-				extraOptions: {
-					command: "update_branch_name",
-					actionName: "Update Branch Name",
-				},
-				query: (args) => args,
-				invalidatesTags: (_r, _e, args) => [
-					invalidatesList(ReduxTag.Stacks), // Probably still needed
-					invalidatesItem(ReduxTag.StackDetails, args.stackId), // This probably is still needed as well
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			removeBranch: build.mutation<
-				void,
-				{
-					projectId: string;
-					stackId?: string;
-					branchName: string;
-				}
-			>({
-				extraOptions: {
-					command: "remove_branch",
-					actionName: "Remove Branch",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesList(ReduxTag.Stacks), // Removing a branch can remove a stack
-					// Removing a branch won't change the sha if the branch is empty
-					invalidatesItem(ReduxTag.StackDetails, args.stackId),
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			reorderStack: build.mutation<
-				void,
-				{ projectId: string; stackId: string; stackOrder: StackOrder }
-			>({
-				extraOptions: {
-					command: "reorder_stack",
-					actionName: "Reorder Stack",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesItem(ReduxTag.StackDetails, args.stackId), // This is probably still needed
-				],
-			}),
-			moveCommit: build.mutation<
-				MoveCommitIllegalAction | null,
-				{ projectId: string; sourceStackId: string; commitId: string; targetStackId: string }
-			>({
-				extraOptions: {
-					command: "move_commit",
-					actionName: "Move Commit",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesList(ReduxTag.WorktreeChanges), // Moving commits can cause conflicts
-					invalidatesItem(ReduxTag.BranchChanges, args.sourceStackId),
-					invalidatesItem(ReduxTag.BranchChanges, args.targetStackId),
-				],
-			}),
-			moveBranch: build.mutation<
-				MoveBranchResult,
-				{
-					projectId: string;
-					sourceStackId: string;
-					subjectBranchName: string;
-					targetStackId: string;
-					targetBranchName: string;
-				}
-			>({
-				extraOptions: {
-					command: "move_branch_legacy",
-					actionName: "Move Branch",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesList(ReduxTag.WorktreeChanges), // Moving commits can cause conflicts
-					invalidatesItem(ReduxTag.BranchChanges, args.sourceStackId),
-					invalidatesItem(ReduxTag.BranchChanges, args.targetStackId),
-				],
-			}),
-			tearOffBranch: build.mutation<
-				MoveBranchResult,
-				{
-					projectId: string;
-					sourceStackId: string;
-					subjectBranchName: string;
-				}
-			>({
-				extraOptions: {
-					command: "tear_off_branch_legacy",
-					actionName: "Tear Off Branch",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => {
-					return [
-						invalidatesList(ReduxTag.HeadSha),
-						invalidatesList(ReduxTag.WorktreeChanges), // Moving commits can cause conflicts
-						invalidatesItem(ReduxTag.BranchChanges, args.sourceStackId), // Affects source stack, new stack is new
-					];
-				},
-			}),
-			integrateUpstreamCommits: build.mutation<
-				void,
-				{
-					projectId: string;
-					stackId: string;
-					seriesName: string;
-					integrationStrategy: SeriesIntegrationStrategy | undefined;
-				}
-			>({
-				extraOptions: {
-					command: "integrate_upstream_commits",
-					actionName: "Integrate Upstream Commits",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesList(ReduxTag.WorktreeChanges),
-					invalidatesItem(ReduxTag.StackDetails, args.stackId),
-					invalidatesItem(ReduxTag.BranchChanges, args.stackId),
-				],
-			}),
-			getInitialIntegrationSteps: build.query<
-				InteractiveIntegrationStep[],
-				{ projectId: string; stackId: string | undefined; branchName: string }
-			>({
-				extraOptions: { command: "get_initial_integration_steps_for_branch" },
-				query: (args) => args,
-				providesTags: (_result, _error, { stackId, branchName }) =>
-					providesItem(ReduxTag.IntegrationSteps, (stackId ?? "--no stack ID--") + branchName),
-			}),
-			integrateBranchWithSteps: build.mutation<
-				void,
-				{
-					projectId: string;
-					stackId: string;
-					branchName: string;
-					steps: InteractiveIntegrationStep[];
-				}
-			>({
-				extraOptions: {
-					command: "integrate_branch_with_steps",
-					actionName: "Integrate Branch with Steps",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesItem(ReduxTag.StackDetails, args.stackId),
-					invalidatesItem(ReduxTag.IntegrationSteps, args.stackId + args.branchName),
-					invalidatesItem(ReduxTag.BranchDetails, args.branchName),
-					invalidatesItem(ReduxTag.BranchChanges, args.stackId),
-				],
-			}),
-			createVirtualBranchFromBranch: build.mutation<
-				CreateBranchFromBranchOutcome,
-				{ projectId: string; branch: string; remote?: string; prNumber?: number }
-			>({
-				extraOptions: {
-					command: "create_virtual_branch_from_branch",
-					actionName: "Create Virtual Branch From Branch",
-				},
-				query: (args) => args,
-				invalidatesTags: [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			deleteLocalBranch: build.mutation<
-				void,
-				{ projectId: string; refname: string; givenName: string }
-			>({
-				extraOptions: {
-					command: "delete_local_branch",
-					actionName: "Delete Local Branch",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, { givenName: branchName }) => [
-					invalidatesItem(ReduxTag.BranchDetails, branchName),
-					providesList(ReduxTag.BranchListing),
-				],
-			}),
-			squashCommits: build.mutation<
-				void,
-				{ projectId: string; stackId: string; sourceCommitIds: string[]; targetCommitId: string }
-			>({
-				extraOptions: {
-					command: "squash_commits",
-					actionName: "Squash Commits",
-				},
-				query: (args) => args,
-				invalidatesTags: [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesList(ReduxTag.WorktreeChanges), // Could cause conflicts
-				],
-			}),
-			newBranchName: build.query<
-				string,
-				{
-					projectId: string;
-				}
-			>({
-				extraOptions: { command: "canned_branch_name" },
-				query: (args) => args,
-			}),
-			normalizeBranchName: build.query<
-				string,
-				{
-					name: string;
-				}
-			>({
-				extraOptions: { command: "normalize_branch_name" },
-				query: (args) => args,
-			}),
-			targetCommits: build.query<
-				EntityState<Commit, string>,
-				{
-					projectId: string;
-					lastCommitId: string | undefined;
-					pageSize: number;
-				}
-			>({
-				extraOptions: { command: "target_commits" },
-				query: (args) => args,
-				transformResponse: (commits: Commit[]) =>
-					commitAdapter.addMany(commitAdapter.getInitialState(), commits),
-			}),
-			splitBranch: build.mutation<
-				{ replacedCommits: [string, string][] },
-				{
-					projectId: string;
-					sourceStackId: string;
-					sourceBranchName: string;
-					newBranchName: string;
-					fileChangesToSplitOff: string[];
-				}
-			>({
-				extraOptions: {
-					command: "split_branch",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesItem(ReduxTag.BranchChanges, args.sourceStackId),
-				],
-			}),
-			splitBranchIntoDependentBranch: build.mutation<
-				{ replacedCommits: [string, string][] },
-				{
-					projectId: string;
-					sourceStackId: string;
-					sourceBranchName: string;
-					newBranchName: string;
-					fileChangesToSplitOff: string[];
-				}
-			>({
-				extraOptions: {
-					command: "split_branch_into_dependent_branch",
-				},
-				query: (args) => args,
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesItem(ReduxTag.BranchChanges, args.sourceStackId),
-				],
-			}),
-			createReference: build.mutation<
-				void,
-				{ projectId: string; stackId: string; request: CreateRefRequest }
-			>({
-				extraOptions: {
-					command: "create_reference",
-					actionName: "Create Reference",
-				},
-				query: (args) => {
-					// TODO: Remove the stack ID from the request args.
-					// The backend doesn't need it, but the frontend does to invalidate the right tags.
-					// We should move away from using the stack ID as the cache key, an move towards some form of branch name instead.
-
-					return { projectId: args.projectId, request: args.request };
-				},
-				invalidatesTags: (_result, _error, args) => [
-					invalidatesItem(ReduxTag.StackDetails, args.stackId), // This is probably still needed. Adding a ref won't change the workspace commit, right?
-				],
-			}),
-			templates: build.query<string[], { projectId: string; forge: string }>({
-				extraOptions: { command: "pr_templates" },
-				query: (args) => args,
-			}),
-			template: build.query<string, { projectId: string; forge: string; relativePath: string }>({
-				extraOptions: { command: "pr_template" },
-				query: (args) => args,
-			}),
-		}),
-	});
-}
-
-const stackAdapter = createEntityAdapter<Stack, string>({
-	selectId: (stack) => stack.id || stack.heads.at(0)?.name || stack.tip,
-});
-const stackSelectors = { ...stackAdapter.getSelectors(), selectNth: createSelectNth<Stack>() };
-
-const commitAdapter = createEntityAdapter<Commit, string>({
-	selectId: (commit) => commit.id,
-});
-const commitSelectors = { ...commitAdapter.getSelectors(), selectNth: createSelectNth<Commit>() };
-
-const upstreamCommitAdapter = createEntityAdapter<UpstreamCommit, string>({
-	selectId: (commit) => commit.id,
-});
-const upstreamCommitSelectors = {
-	...upstreamCommitAdapter.getSelectors(),
-	selectNth: createSelectNth<UpstreamCommit>(),
-};
-
-const changesAdapter = createEntityAdapter<TreeChange, string>({
-	selectId: (change) => change.path,
-});
-
-const changesSelectors = changesAdapter.getSelectors();
-
-const selectChangesByPaths = createSelectByIds<TreeChange>();
-
-const branchDetailsAdapter = createEntityAdapter<BranchDetails, string>({
-	selectId: (branch) => branch.name,
-});
-
-const branchDetailsSelectors = branchDetailsAdapter.getSelectors();
