+++ apps/desktop/src/lib/upstream/upstreamIntegrationService.svelte.ts
@@ -1,37 +1,25 @@
-import { invalidatesList, providesList, ReduxTag } from "$lib/state/tags";
 import { InjectionToken } from "@gitbutler/core/context";
 import { isDefined } from "@gitbutler/ui/utils/typeguards";
 import type { StackService } from "$lib/stacks/stackService.svelte";
 import type { BackendApi } from "$lib/state/clientState.svelte";
-import type {
-	BaseBranchResolution,
-	BaseBranchResolutionApproach,
-	BranchStatusesResponse,
-	IntegrationOutcome,
-	Resolution,
-	StackStatusesWithBranchesV3,
-} from "$lib/upstream/types";
+import type { StackStatusesWithBranchesV3 } from "$lib/upstream/types";
 
 export const UPSTREAM_INTEGRATION_SERVICE = new InjectionToken<UpstreamIntegrationService>(
 	"UpstreamIntegrationService",
 );
 
 export class UpstreamIntegrationService {
-	private api: ReturnType<typeof injectEndpoints>;
-
 	constructor(
-		api: BackendApi,
+		private backendApi: BackendApi,
 		private stackService: StackService,
-	) {
-		this.api = injectEndpoints(api);
-	}
+	) {}
 
 	async upstreamStatuses(
 		projectId: string,
 		targetCommitOid: string | undefined,
 	): Promise<StackStatusesWithBranchesV3 | undefined> {
 		const stacks = await this.stackService.fetchStacks(projectId);
-		const branchStatuses = await this.api.endpoints.upstreamIntegrationStatuses.fetch({
+		const branchStatuses = await this.backendApi.endpoints.upstreamIntegrationStatuses.fetch({
 			projectId,
 			targetCommitOid,
 		});
@@ -58,59 +46,14 @@ export class UpstreamIntegrationService {
 	}
 
 	resolveUpstreamIntegration() {
-		return this.api.endpoints.resolveUpstreamIntegration.useMutation();
+		return this.backendApi.endpoints.resolveUpstreamIntegration.useMutation();
 	}
 
 	get resolveUpstreamIntegrationMutation() {
-		return this.api.endpoints.resolveUpstreamIntegration.mutate;
+		return this.backendApi.endpoints.resolveUpstreamIntegration.mutate;
 	}
 
 	integrateUpstream() {
-		return this.api.endpoints.integrateUpstream.useMutation();
+		return this.backendApi.endpoints.integrateUpstream.useMutation();
 	}
-}
-
-function injectEndpoints(api: BackendApi) {
-	return api.injectEndpoints({
-		endpoints: (build) => ({
-			upstreamIntegrationStatuses: build.query<
-				BranchStatusesResponse,
-				{ projectId: string; targetCommitOid?: string }
-			>({
-				extraOptions: { command: "upstream_integration_statuses" },
-				query: (args) => args,
-				providesTags: [providesList(ReduxTag.UpstreamIntegrationStatus)],
-			}),
-			integrateUpstream: build.mutation<
-				IntegrationOutcome,
-				{
-					projectId: string;
-					resolutions: Resolution[];
-					baseBranchResolution?: BaseBranchResolution;
-				}
-			>({
-				extraOptions: {
-					command: "integrate_upstream",
-					actionName: "Integrate Upstream",
-				},
-				query: (args) => args,
-				invalidatesTags: [
-					invalidatesList(ReduxTag.UpstreamIntegrationStatus),
-					invalidatesList(ReduxTag.HeadSha),
-					invalidatesList(ReduxTag.BranchListing),
-				],
-			}),
-			resolveUpstreamIntegration: build.mutation<
-				string,
-				{ projectId: string; resolutionApproach: { type: BaseBranchResolutionApproach } }
-			>({
-				extraOptions: {
-					command: `resolve_upstream_integration`,
-					actionName: "Resolve Integrate Upstream",
-				},
-				query: (args) => args,
-				invalidatesTags: [invalidatesList(ReduxTag.UpstreamIntegrationStatus)],
-			}),
-		}),
-	});
 }
